backend/pyproject.toml
[tool.poetry]
name = "django-sveltekit-backend"
version = "0.1.0"
description = "Django backend application"
authors = []
readme = "README.md"
packages = []
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
asyncer = "^0.0.8"
backoff = "*"
beautifulsoup4 = "*"
celery = "^5.4.0"
cloudpickle = "^3.1.1"
dj-database-url = "*"
dj-rest-auth = {extras = ["with_social"], version = "*"}
django = {extras = ["argon2"], version = "~=4.2"}
django-anymail = {extras = ["mailgun"], version = "*"}
django-celery-beat = "^2.7.0"
django-cors-headers = "*"
django-csp = "*"
django-debug-toolbar = "*"
django-extensions = "*"
django-filter = "*"
django-nested-admin = "*"
django-permissions-policy = "*"
django-revproxy = "*"
django-rich = "*"
django-storages = {extras = ["boto3"], version = "*"}
djangorestframework = "*"
djangorestframework-api-key = "3.*"
djangorestframework-simplejwt = "*"
drf-spectacular = "*"
dspy-ai = "*"
flower = "*"
google-api-python-client = "*"
google-auth = "*"
google-cloud-run = "*"
groq = "*"
gunicorn = "*"
httpx = "==0.27.2"
hypothesis = "*"
json-repair = "^0.39.1"
langchain = "*"
langchain-openai = "*"
markdownify = "*"
multion = "*"
nest_asyncio = "*"
numpy = "*"
open-interpreter = "*"
opencv-python = "*"
pillow = "*"
psycopg2-binary = "*"
pypdf = "*"
pytesseract = "*"
python-dotenv = "*"
rapidfuzz = "*"
redis = "^5.2.1"
sentry-sdk = "*"
weasyprint = "*"
whitenoise = {extras = ["brotli"], version = "*"}

[tool.poetry.group.dev.dependencies]
black = "*"
django-types = "*"
djangorestframework-types = "==0.7.1"
hypothesis = "*"
inotify = "*"
ipython = "*"
isort = "*"
model-bakery = "*"
pre-commit = "*"
ptpython = "*"
pyright = "^1.1.401"
pytest = "*"
pytest-asyncio = "==0.21.1"
pytest-cov = "*"
pytest-django = "*"
pytest-mock = "*"
pytest-structlog = "*"
pytest-xdist = "*"
reorder-python-imports = "*"
requests-mock = "*"
respx = "*"
ruff = "*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.coverage.run]
branch = true
omit = ["manage.py", "project/asgi.py", "project/wsgi.py"]
source = ["."]

[tool.coverage.report]
exclude_lines = [
    "class .*\\bProtocol\\):",
    "@overload",
    "@typing.overload",
    "pragma: no cover",
]
fail_under = 0

[tool.django-stubs]
django_settings_module = "project.settings"

[tool.pyright]
stubPath = "./typings"
reportImportCycles = false
typeCheckingMode = "basic"
venv = ".venv"
venvPath = "."
# ignore migrations, and also the files pyright excludes by default.
exclude = ["**/migrations", "**/node_modules", "**/__pycache__", "**/.*"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "project.settings"
addopts = "--cov --dist=no -n 0 --reuse-db"
python_files = ["tests.py", "**/tests/test_*.py", "**/tests/**/test_*.py"]

[tool.ruff]
# Ignore max line length - black will handle this and it's okay to go over sometimes.
ignore = ["E501"]

[tool.isort]
profile = "black"
